Game service

+++ frontend/src/api/game-service/games/game.ts
@@ -3,8 +3,9 @@ type Success = { ok: true; message: string };
 
 type SimpleResult = Success | Failure;
 
+// When a game is created or launched, it gives its info, including maxScore and if it's part of a tournament
 type GameInfoResult =
-    | { ok: true; gameId: number, status: string, player_a: string, player_b: string }
+    | { ok: true; gameId: number, status: string, player_a: string, player_b: string, maxScore: number, tournament_id: number }
     | Failure
 
 type PendingGamesInfos = {
@@ -36,14 +37,18 @@ export type AvailableGamesResult = AvailableGamesList | Failure;
 type AllGamesList = { ok: true; games: AllGamesInfos[] }
 export type AllGamesResult = AllGamesList | Failure;
 
-// POST /games/game request creates a new game for the user, taking names for players
-export async function createLocalGame(userId: number, player_a: string, player_b: string): Promise<GameInfoResult> {
-    // TODO = Log check not functional
+// All routes going to game service begin with : "/api/games/"
+
+// POST /game
+// Creates a new game for the user, taking names for players
+// Security : Accessible for every logged-in user
+// TODO = Add AI registration
+export async function createLocalGame(player_a: string, player_b: string, maxScore: number = 7): Promise<GameInfoResult> {
     const token = localStorage.getItem("token");
     if (!token)
         return { ok: false, error: "No token"};
 
-    console.log('userId = ' + userId + ' playA ' + player_a + ' playB ' + player_b);
+    console.log(' playA ' + player_a + ' playB ' + player_b);
     const res = await fetch('/api/games/game', {
         method: 'POST',
         headers: {
@@ -51,9 +56,9 @@ export async function createLocalGame(userId: number, player_a: string, player_b
             'Authorization': `Bearer ${token}`
         },
         body: JSON.stringify({
-            userId: userId,
             player_a: player_a,
             player_b: player_b,
+			maxScore: maxScore
         })
     });
     const data = await res.json();
@@ -65,23 +70,28 @@ export async function createLocalGame(userId: number, player_a: string, player_b
             status: data.status,
             player_a: data.player_a,
             player_b: data.player_b,
+			tournament_id: data.tournament,
+			maxScore: data.maxScore
         };
     } else {
         // Invalid or expired token = Disconnect
-        alert("API Error starting game : " + data?.error as string  || "Game start impossible");
+        alert("❌ API Error starting game : " + data?.error as string  || "Game start impossible");
         return { ok: false, error: data?.error as string  || "Game start impossible" };
     }
 }
 
-// POST /:game_id to launch a game and reserve game server in backend
+// POST /:game_id
 // Response: {
 //   game_id: number,
 //   status: "ongoing",
 //   players: [string, string]
 // }
+// Launch a game and reserve a game server in backend
+// Security : is gonna be possible only if the logges-in user is the owner of the game
 export async function startGame(gameId: number): Promise<GameInfoResult> {
-    // TODO = Add log check
-
+	const token = localStorage.getItem("token");
+    if (!token)
+        return { ok: false, error: "No token"};
     if (!gameId) {
         return { ok: false, error: "No game ID given" };
     }
@@ -89,22 +99,31 @@ export async function startGame(gameId: number): Promise<GameInfoResult> {
         const request = await fetch(`/api/games/${gameId}`, {
             method: 'POST',
             headers: {
-                // 'Content-Type': 'application/json',
-                // Auth header
+				// 'Content-Type': 'application/json', // Useless because no body provided
+                'Authorization': `Bearer ${token}`
             }
         });
         if (!request.ok) {
             throw new Error('Unable to start game ' + request.status);
         }
         const data = await request.json();
-        return { ok: true, gameId: data.game_id, status: data.status, player_a: data.player_a, player_b: data.player_b };
+        return { ok: true, 
+			gameId: data.game_id, 
+			status: data.status, 
+			player_a: data.player_a, 
+			player_b: data.player_b, 
+			tournament_id: data.tournament_id, 
+			maxScore: data.maxScore 
+		};
     }
     catch (error) {
         return { ok: false, error: error as string  };
     }
 }
 
-// GET /:userId/games = all available PENDING games for a user, gives only useful infos
+// GET /:userId/games
+// All available PENDING non-tournament games for a user, gives only useful infos
+// Security : Accessible for every logged-in user
 export async function getAvailableGames(userId: number): Promise<AvailableGamesResult> {
     try {
         const allGamesResult = await getAllGames(userId);
@@ -113,6 +132,7 @@ export async function getAvailableGames(userId: number): Promise<AvailableGamesR
         }
         const pendingGames: PendingGamesInfos[] = allGamesResult.games
             .filter(game => game.status === 'pending')
+			.filter(game => game.tournament_id == 0)
             .map(game => ({
                 id: game.id,
                 status: game.status,
@@ -124,15 +144,46 @@ export async function getAvailableGames(userId: number): Promise<AvailableGamesR
         return { ok: true, games: pendingGames };
 
     } catch (error) {
-        console.error('error filtering pending games', error as string );
+        console.error('❌ Error filtering pending games', error as string );
         return { ok: false, error: error as string  };
     }
 }
 
-// GET /:userId/games = all games for a user (useful for history, gives you every info available on each game)
-export async function getAllGames(userId: number): Promise<AllGamesResult> {
-    // TODO = Add log check
+// GET /:userId/games
+// All available FINISHED games for a user, gives only useful infos
+// Security : Accessible for every logged-in user
+export async function getFinishedGames(userId: number): Promise<AvailableGamesResult> {
+    try {
+        const allGamesResult = await getAllGames(userId);
+        if (!allGamesResult.ok) {
+            return { ok: false, error: allGamesResult.error };
+        }
+        const pendingGames: PendingGamesInfos[] = allGamesResult.games
+            .filter(game => game.status === 'finished')
+            .map(game => ({
+                id: game.id,
+                status: game.status,
+                player_a: game.player_a,
+                player_b: game.player_b,
+                created_at: game.created_at
+            }));
 
+        return { ok: true, games: pendingGames };
+
+    } catch (error) {
+        console.error('❌ Error filtering finished games', error as string );
+        return { ok: false, error: error as string  };
+    }
+}
+
+
+// GET /:userId/games
+// Gives all games for a user (useful for history, gives you every info available on each game)
+// Security : Accessible for every logged-in user
+export async function getAllGames(userId: number): Promise<AllGamesResult> {
+    const token = localStorage.getItem("token");
+    if (!token)
+        return { ok: false, error: "No token"};
     if (!userId) {
         return { ok: false, error: 'User ID is required' };
     }
@@ -140,8 +191,8 @@ export async function getAllGames(userId: number): Promise<AllGamesResult> {
         const response = await fetch(`/api/games/${userId}/games`, {
             method: 'GET',
             headers: {
-                // 'Content-Type': 'application/json',
-                // Auth header
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
             }
         });
 
@@ -154,24 +205,33 @@ export async function getAllGames(userId: number): Promise<AllGamesResult> {
         return { ok: true, games: games };
 
     } catch (error) {
-        console.error('Error fetching available games:', error);
+        console.error('❌ Error fetching available games:', error);
         return { ok: false, error: error as string  };
     }
 }
 
+// DELETE /:gameId
+// Delete a game in backend
+// TODO : Works but in case of an error, writes alert("Error: Error:...")
+// Security : is gonna be possible only if the logges-in user is the owner of the game
 export async function deleteGame(gameId: number): Promise<SimpleResult> {
-    // TODO = Add log check
+    const token = localStorage.getItem("token");
+    if (!token)
+        return { ok: false, error: "No token"};
+
     if (!gameId)
         return {ok: false, error: 'gameId required'};
     try {
         const response = await fetch(`/api/games/${gameId}`, {
             method: 'DELETE',
             headers:  {
-                // Auth header
+				// 'Content-Type': 'application/json', // Useless because no body provided
+                'Authorization': `Bearer ${token}`
             }
         });
         if (!response.ok) {
-            throw new Error(`Unable to delete game because ` + response.status);
+			const data = await response.json();
+            throw new Error(`Unable to delete game ->` + data.error);
         }
         return { ok: true, message: gameId + ' game has been deleted' };
     } catch(error) {
